# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/afficher_individus/":{"get":{"summary":"Read Individus","operationId":"read_individus_afficher_individus__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/IndividuRead"},"type":"array","title":"Response Read Individus Afficher Individus  Get"}}}}}}},"/ajouter_individu/":{"post":{"summary":"Create Individu","operationId":"create_individu_ajouter_individu__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndividuCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/update_individu/{id_individu}":{"put":{"summary":"Update Individu","operationId":"update_individu_update_individu__id_individu__put","parameters":[{"name":"id_individu","in":"path","required":true,"schema":{"type":"integer","title":"Id Individu"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndividuUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/update_relation/parent_enfant/":{"put":{"summary":"Update Relation Parent Enfant","operationId":"update_relation_parent_enfant_update_relation_parent_enfant__put","parameters":[{"name":"old_parent","in":"query","required":true,"schema":{"type":"integer","title":"Old Parent"}},{"name":"old_enfant","in":"query","required":true,"schema":{"type":"integer","title":"Old Enfant"}},{"name":"old_id_relation","in":"query","required":true,"schema":{"type":"integer","title":"Old Id Relation"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelationParentEnfantUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/update_relation/couple/":{"put":{"summary":"Update Relation Couple","operationId":"update_relation_couple_update_relation_couple__put","parameters":[{"name":"old_individu1","in":"query","required":true,"schema":{"type":"integer","title":"Old Individu1"}},{"name":"old_individu2","in":"query","required":true,"schema":{"type":"integer","title":"Old Individu2"}},{"name":"old_id_relation","in":"query","required":true,"schema":{"type":"integer","title":"Old Id Relation"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelationCoupleUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/types_relation/":{"get":{"summary":"Get Types Relation","operationId":"get_types_relation_types_relation__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/delete_individu/{id_individu}":{"delete":{"summary":"Delete Individu","operationId":"delete_individu_delete_individu__id_individu__delete","parameters":[{"name":"id_individu","in":"path","required":true,"schema":{"type":"integer","title":"Id Individu"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete_relation_parent_enfant/":{"delete":{"summary":"Delete Relation Parent Enfant","operationId":"delete_relation_parent_enfant_delete_relation_parent_enfant__delete","parameters":[{"name":"id_parent","in":"query","required":true,"schema":{"type":"integer","title":"Id Parent"}},{"name":"id_enfant","in":"query","required":true,"schema":{"type":"integer","title":"Id Enfant"}},{"name":"id_relation","in":"query","required":true,"schema":{"type":"integer","title":"Id Relation"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete_relation_couple/":{"delete":{"summary":"Delete Relation Couple","operationId":"delete_relation_couple_delete_relation_couple__delete","parameters":[{"name":"id_individu1","in":"query","required":true,"schema":{"type":"integer","title":"Id Individu1"}},{"name":"id_individu2","in":"query","required":true,"schema":{"type":"integer","title":"Id Individu2"}},{"name":"id_relation","in":"query","required":true,"schema":{"type":"integer","title":"Id Relation"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ajouter_relation/parent_enfant/":{"post":{"summary":"Add Relation Parent Enfant","operationId":"add_relation_parent_enfant_ajouter_relation_parent_enfant__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelationParentEnfantCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ajouter_relation/couple/":{"post":{"summary":"Add Relation Couple","operationId":"add_relation_couple_ajouter_relation_couple__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelationCoupleCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete_nom_individu/":{"delete":{"summary":"Delete Nom Individu","operationId":"delete_nom_individu_delete_nom_individu__delete","parameters":[{"name":"id_individu","in":"query","required":true,"schema":{"type":"integer","title":"Id Individu"}},{"name":"nom","in":"query","required":true,"schema":{"type":"string","title":"Nom"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete_prenom_individu/":{"delete":{"summary":"Delete Prenom Individu","operationId":"delete_prenom_individu_delete_prenom_individu__delete","parameters":[{"name":"id_individu","in":"query","required":true,"schema":{"type":"integer","title":"Id Individu"}},{"name":"prenom","in":"query","required":true,"schema":{"type":"string","title":"Prenom"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IndividuCreate":{"properties":{"date_naissance":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Naissance"},"date_deces":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Deces"},"noms":{"items":{"type":"string"},"type":"array","title":"Noms"},"prenoms":{"items":{"type":"string"},"type":"array","title":"Prenoms"}},"type":"object","required":["noms","prenoms"],"title":"IndividuCreate"},"IndividuRead":{"properties":{"id_individu":{"type":"integer","title":"Id Individu"},"date_naissance":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Naissance"},"date_deces":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Deces"},"noms":{"items":{"type":"string"},"type":"array","title":"Noms","default":[]},"prenoms":{"items":{"type":"string"},"type":"array","title":"Prenoms","default":[]}},"type":"object","required":["id_individu"],"title":"IndividuRead"},"IndividuUpdate":{"properties":{"date_naissance":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Naissance"},"date_deces":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Deces"},"nom":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nom"},"prenoms":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Prenoms"}},"type":"object","title":"IndividuUpdate"},"RelationCoupleCreate":{"properties":{"id_individu1":{"type":"integer","title":"Id Individu1"},"id_individu2":{"type":"integer","title":"Id Individu2"},"id_relation":{"type":"integer","title":"Id Relation"}},"type":"object","required":["id_individu1","id_individu2","id_relation"],"title":"RelationCoupleCreate"},"RelationCoupleUpdate":{"properties":{"id_individu1":{"type":"integer","title":"Id Individu1"},"id_individu2":{"type":"integer","title":"Id Individu2"},"id_relation":{"type":"integer","title":"Id Relation"}},"type":"object","required":["id_individu1","id_individu2","id_relation"],"title":"RelationCoupleUpdate"},"RelationParentEnfantCreate":{"properties":{"id_parent":{"type":"integer","title":"Id Parent"},"id_enfant":{"type":"integer","title":"Id Enfant"},"id_relation":{"type":"integer","title":"Id Relation"}},"type":"object","required":["id_parent","id_enfant","id_relation"],"title":"RelationParentEnfantCreate"},"RelationParentEnfantUpdate":{"properties":{"id_parent":{"type":"integer","title":"Id Parent"},"id_enfant":{"type":"integer","title":"Id Enfant"},"id_relation":{"type":"integer","title":"Id Relation"}},"type":"object","required":["id_parent","id_enfant","id_relation"],"title":"RelationParentEnfantUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}